Add Parser.java. Parser.java encapsulates methods that deal with making sense of user command
